Added to typeClass amendmentment breakdownRow
